+++ pipelines/red_balls_15ft.vpr
@@ -1,0 +1,58 @@
+area_max:3.8862602496000003
+area_min:0
+area_similarity:0
+aspect_max:2.5919999999999996
+aspect_min:0
+black_level:0
+blue_balance:2500
+calibration_type:0
+contour_grouping:0
+contour_sort_final:0
+convexity_max:100
+convexity_min:29.8
+corner_approx:2
+cross_a_a:1
+cross_a_x:0
+cross_a_y:0
+cross_b_a:1
+cross_b_x:0
+cross_b_y:0
+desc:red_balls_15ft
+desired_contour_region:0
+dilation_steps:0
+direction_filter:0
+dual_close_sort_origin:0
+erosion_steps:0
+exposure:35
+force_convex:1
+hue_max:15
+hue_min:0
+image_flip:0
+image_source:0
+img_to_show:0
+intersection_filter:0
+pipeline_led_enabled:1
+pipeline_led_power:0
+pipeline_res:0
+pipeline_type:0
+red_balance:862
+roi_x:0.000000
+roi_y:0.000000
+sat_max:255
+sat_min:100
+send_corners:0
+send_raw_contours:0
+solve3d:0
+solve3d_algo:0
+solve3d_bindtarget:1
+solve3d_conf:0.990000
+solve3d_error:8
+solve3d_guess:0
+solve3d_iterations:50
+solve3d_precies:1
+solve3d_precise:0
+solve3d_zoffset:0
+val_max:255
+val_min:134
+y_max:1.000000
+y_min:-1.000000
