+++ src/main/java/frc/robot/commands/Deprecated/MoveConveyor.java
@@ -1,17 +1,15 @@
-package frc.robot.commands;
+package frc.robot.commands.Deprecated;
 
-import edu.wpi.first.wpilibj2.command.InstantCommand;
+import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.subsystems.ConveyorSubsystem;
-import frc.robot.subsystems.IntakeSubsystem;
 import frc.robot.Constants.*;
 
-public class MoveConveyor extends InstantCommand {
-    private IntakeSubsystem intakeSubsystem;
+public class MoveConveyor extends CommandBase {
+    // private IntakeSubsystem intakeSubsystem;
     private ConveyorSubsystem conveyorSubsystem;
     
     public MoveConveyor(ConveyorSubsystem conveyorSubsystem){
         this.conveyorSubsystem = conveyorSubsystem;
-      
         addRequirements(conveyorSubsystem);
     }
     @Override
