+++ src/main/java/frc/robot/commands/Deprecated/ShooterOff.java
@@ -1,4 +1,4 @@
-package frc.robot.commands;
+package frc.robot.commands.Deprecated;
 
 import edu.wpi.first.wpilibj2.command.InstantCommand;
 import frc.robot.subsystems.ShooterSubsystem;
