+++ src/main/java/frc/robot/commands/Deprecated/StopConveyor.java
@@ -1,12 +1,9 @@
-package frc.robot.commands;
+package frc.robot.commands.Deprecated;
 
 import edu.wpi.first.wpilibj2.command.InstantCommand;
 import frc.robot.subsystems.ConveyorSubsystem;
-import frc.robot.subsystems.IntakeSubsystem;
-import frc.robot.Constants.*;
 
 public class StopConveyor extends InstantCommand {
-    private IntakeSubsystem intakeSubsystem;
     private ConveyorSubsystem conveyorSubsystem;
     
     public StopConveyor(ConveyorSubsystem conveyorSubsystem){
