+++ src/main/java/frc/robot/commands/SpitOutBall.java
@@ -1,14 +1,13 @@
 package frc.robot.commands;
-import edu.wpi.first.wpilibj2.command.InstantCommand;
+import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.subsystems.ConveyorSubsystem;
 import frc.robot.subsystems.IntakeSubsystem;
-import frc.robot.Constants;
 import frc.robot.Constants.ConveyorConstants;
 import frc.robot.Constants.IntakeConstants;
-public class SpitOutBall extends InstantCommand{
+
+public class SpitOutBall extends CommandBase{
     private IntakeSubsystem  intakeSubsystem;
     private ConveyorSubsystem conveyorSubsystem;
-    double spitOutBall = Constants.IntakeConstants.SPIT_OUT_BALL;
 
     public SpitOutBall(IntakeSubsystem intakeSubsystem, ConveyorSubsystem conveyorSubsystem){
         this.intakeSubsystem = intakeSubsystem;
@@ -17,8 +16,13 @@ public class SpitOutBall extends InstantCommand{
     }
     @Override
     public void initialize(){
-        //reverse intake and conveyor at same time
-        //intakeSubsystem.reverseIntake(spitOutBall);
+        intakeSubsystem.reverseIntake(IntakeConstants.SPIT_OUT_BALL);
         conveyorSubsystem.moveConveyor(ConveyorConstants.CONVEYOR_SPEED * -1);
+    }
+
+    @Override
+    public void end(boolean interrupted) {
+      conveyorSubsystem.stopConveyor();
+      intakeSubsystem.stopIntake();
     }
 }
