+++ src/main/java/frc/robot/commands/TakeInBall.java
@@ -1,11 +1,11 @@
 package frc.robot.commands;
 
-import edu.wpi.first.wpilibj2.command.InstantCommand;
+import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.subsystems.ConveyorSubsystem;
 import frc.robot.subsystems.IntakeSubsystem;
 import frc.robot.Constants.*;
 
-public class TakeInBall extends InstantCommand {
+public class TakeInBall extends CommandBase {
     private ConveyorSubsystem conveyorSubsystem;
     private IntakeSubsystem intakeSubsystem;
